Extract shared Exceptional fields into a BaseExceptional class.
